+++ Assets/LDtkUnity/CHANGELOG.md
@@ -1,9 +1,10 @@
 # 2.0.4
 ###### June XX, 2021
-- Added a help reference button for each importer section 
+- Added a toggle in the importer to create a CompositeCollider2D component on all IntGrid tilemaps. 
 - Added support for LDtk layer pixel offsets
-- Changed how the level borders are drawn in scene view
+- Added a help reference button for each importer section 
 - Added extra height to `MultiLine` fields in the LDtkFields component
+- Changed how the level borders are drawn in scene view
 - Other minor tweaks and fixes
 
 # 2.0.3