+++ Assets/LDtkUnity/CHANGELOG.md
@@ -1,3 +1,6 @@
+# 2.0.5
+###### June XX, 2021
+
 # 2.0.4
 ###### June 6, 2021
 - Added a toggle in the importer to create a CompositeCollider2D component on all IntGrid tilemaps